+++ test-curves/src/bls12_381/g1.rs
@@ -1,7 +1,7 @@
 use crate::bls12_381::*;
 use ark_ec::{
-    models::{ModelParameters, SWModelParameters},
-    short_weierstrass::*,
+    models::CurveConfig,
+    short_weierstrass::{self, *},
 };
 use ark_ff::{MontFp, Zero};
 
@@ -11,7 +11,7 @@ pub type G1Projective = Projective<Parameters>;
 #[derive(Clone, Default, PartialEq, Eq)]
 pub struct Parameters;
 
-impl ModelParameters for Parameters {
+impl CurveConfig for Parameters {
     type BaseField = Fq;
     type ScalarField = Fr;
 
@@ -24,7 +24,7 @@ impl ModelParameters for Parameters {
     const COFACTOR_INV: Fr = MontFp!(Fr, "52435875175126190458656871551744051925719901746859129887267498875565241663483");
 }
 
-impl SWModelParameters for Parameters {
+impl short_weierstrass::SWCurveConfig for Parameters {
     /// COEFF_A = 0
     const COEFF_A: Fq = MontFp!(Fq, "0");